+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Upgrading to 3.8.x versions requires that you upgrade to latest 3.5.x version fi
 - Fix permanently remove deleted emails admin action, #677, #676
 - Set max-width to inline markdown images, #679
 - Fix saving custom field attached to multiple projects, #683, #682
+- Use `EventManager::dispatch()` to get consistent API from Eventum itself, #678
 
 [3.8.3]: https://github.com/eventum/eventum/compare/v3.8.2...master